Create 数字根.cpp
（1）Problem Description
In ancient Indian mathematics, a Vedic square is a variation on a typical 9×9 multiplication table. The entry in each cell is the digital root of the product of the column and row headings.
（2）Vedic Star
By replacing a specific number by asterisk whereas the others by space within the Vedic square, you can find some distinct shapes each with some form of reflection symmetry.